requirements.txt
Flask==2.0.2
PyJWT==2.3.0
mysql-connector-python==8.0.28
